# Eval_U3A_TejelanasVivi/docs/swagger.json
{"openapi":"3.0.0","info":{"title":"API Tejelanas","version":"1.0.0","description":"Documentación del CRUD de productos para Tejelanas Vivi"},"servers":[{"url":"http://localhost:3000"}],"paths":{"/api/productos":{"get":{"summary":"Obtener todos los productos","description":"Este endpoint devuelve todos los productos disponibles.\n🔄 **Caché:** Los resultados se almacenan durante 60 segundos para mejorar el rendimiento.\n🚫 **Rate limiting:** Limitado a 100 solicitudes por minuto por IP.\n⏱️ **Tiempo de respuesta esperado:** entre 10 ms y 50 ms en condiciones normales.\n🧠 **Consejo de optimización:** Evite enviar múltiples solicitudes seguidas si los datos no cambian frecuentemente. El uso de cache ayuda a reducir la carga del servidor.\n","responses":{"200":{"description":"Lista de productos","content":{"application/json":{"example":[{"id":1,"nombre":"Lana Premium","descripcion":"Suave y cálida","precio":9500,"stock":10}]}}}}},"post":{"summary":"Crear un nuevo producto","description":"Este endpoint permite crear un nuevo producto en la base de datos.\n🚫 **Rate limiting:** Limitado a 100 solicitudes por minuto por IP.\n⏱️ **Tiempo de respuesta esperado:** 15 a 80 ms.\n🧠 **Consejo:** Valide los datos antes de enviarlos para evitar errores y ahorrar tráfico.\n","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"precio":{"type":"number"},"stock":{"type":"integer"}}},"example":{"nombre":"Lana Merino","descripcion":"Color natural, ideal para invierno","precio":8900,"stock":15}}}},"responses":{"201":{"description":"Producto creado"}}}},"/api/productos/{id}":{"get":{"summary":"Obtener un producto por ID","description":"Este endpoint devuelve los detalles de un producto según su ID.\n🚫 **Rate limiting:** Limitado a 100 solicitudes por minuto por IP.\n⏱️ **Tiempo de respuesta esperado:** 10 a 40 ms.\n🧠 **Consejo:** Asegúrese de que el ID sea válido para evitar errores 404.\n","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Un producto","content":{"application/json":{"example":{"id":2,"nombre":"Algodón Natural","descripcion":"100% algodón","precio":7800,"stock":5}}}},"404":{"description":"Producto no encontrado"}}},"put":{"summary":"Actualizar un producto existente","description":"Este endpoint permite modificar los datos de un producto por su ID.\n🚫 **Rate limiting:** Limitado a 100 solicitudes por minuto por IP.\n⏱️ **Tiempo de respuesta esperado:** 15 a 100 ms.\n🧠 **Consejo:** Asegúrese de que el ID exista y de enviar todos los campos correctamente para evitar fallos.\n","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"},"description":"ID del producto a actualizar"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"precio":{"type":"number"},"stock":{"type":"integer"}}},"example":{"nombre":"Lana actualizada","descripcion":"Cambio de descripción","precio":9200,"stock":20}}}},"responses":{"200":{"description":"Producto actualizado"},"404":{"description":"Producto no encontrado"}}},"delete":{"summary":"Eliminar un producto","description":"Este endpoint elimina un producto por su ID.\n🚫 **Rate limiting:** Limitado a 100 solicitudes por minuto por IP.\n⏱️ **Tiempo de respuesta esperado:** 10 a 50 ms.\n🧠 **Consejo:** Confirme que el ID sea válido antes de eliminar.\n","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"},"description":"ID del producto a eliminar"}],"responses":{"200":{"description":"Producto eliminado"},"404":{"description":"Producto no encontrado"}}}}},"components":{},"tags":[]}